BackEnd/HEXA_V1/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.5
packaging==24.2
PyJWT==2.9.0
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2